requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.3
fonttools==4.60.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lab==8.5
lark==1.2.2
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.1.4
multidict==6.7.1
mysql-connector-python==9.6.0
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.3
openai==2.30.0
packaging==25.0
panda==0.3.1
pandas==2.3.2
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
plotly==6.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
protobuf==6.32.1
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==25.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
queuelib==1.8.0
referencing==0.36.2
requests==2.32.5
requests-file==3.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
scikit-learn==1.8.0
scipy==1.16.3
Scrapy==2.13.3
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
streamlit==1.49.1
streamlit-autorefresh==1.0.1
stripe==15.0.1
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.3
traitlets==5.14.3
twilio==9.10.4
Twisted==25.5.0
txt2tags==3.9
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.7
widgetsnbextension==4.0.14
yarl==1.23.0
zope.interface==8.0.1
